+++ app/src/main/java/nextstep/jwp/ui/request/HttpRequest.java
@@ -1,5 +1,6 @@
 package nextstep.jwp.ui.request;
 
+import nextstep.jwp.ui.RequestHandler;
 import nextstep.jwp.ui.common.HttpHeaders;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
